FAC-569:feature:Impl GetNotificationPlatformSettingsService
