--- flickit-assessment-users/src/main/java/org/flickit/assessment/users/adapter/in/rest/user/GetUserSubscriberHashResponseDto.java
@@ -1,4 +1,0 @@
-package org.flickit.assessment.users.adapter.in.rest.user;
-
-public record GetUserSubscriberHashResponseDto(String subscriberHash) {
-}
