+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/service/user/GetNotificationPlatformSettingsService.java
@@ -1,0 +1,18 @@
+package org.flickit.assessment.users.application.service.user;
+
+import lombok.RequiredArgsConstructor;
+import org.flickit.assessment.common.application.domain.notification.NotificationSenderSettingProvider;
+import org.flickit.assessment.users.application.port.in.user.GetNotificationPlatformSettingsUseCase;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class GetNotificationPlatformSettingsService implements GetNotificationPlatformSettingsUseCase {
+
+    private final NotificationSenderSettingProvider settingProvider;
+
+    @Override
+    public Result getNotificationPlatformSettings(Param param) {
+        return new Result(settingProvider.getSettings(param.getCurrentUserId()));
+    }
+}
